requirements.txt
django>=1.11.0,<1.11.99
django-flat-responsive
djangorestframework==3.3
pytz
requests
Adafruit-DHT==1.3.0
RPi.GPIO